introduce mapArgX(WithRepresentation)
as complement of withArgX which does not allow to change the arg type.
Next logic step would be to provide a `replace` which does not expect
a lambda but just a value and representation. But I will first see
if I need this often enough so that an own method is justified.