ArgsSourceLike, ArgsSourceOptions and fix splitting of Tuple
Unfortunately, it seems like @Language annotations are ignored by
Intellij if they appear in dependencies. Hence, search for annotations
which itself are annotated with ArgsSourceLike and extract the
methodName from there. Due to this, move the remaining properties into
an own annotation ArgsSourceOptions which shall also be usable on
a class (or super-class). Introduce AnnotationData and
AnnotationDataDeducer which puts the annotation properties into
AnnotationData

moreover:
- remove Repeatable from ArgsSource, one can concatenate
  ArgsGenerators
- don't use computeIf in DefaultComponentFactoryContainer use
  putIfAbsent (as it was in Atrium), otherwise we can run into
  recursive update exceptions in case of dependencies.